+++ include/html/grids.php
@@ -51,7 +51,7 @@ function getHistory($cfg_file, $accordion_id, $open_first_history_tab = false) {
    <li><a data-toggle="tab" href="#menu2"><span class="glyphicon glyphicon-king" aria-hidden="true"></span> Web Admins</a></li>
    <li><a data-toggle="tab" href="#menu3"><span class="glyphicon glyphicon-edit" aria-hidden="true"></span> Configs</a></li>
 </ul>
-<div class="tab-content">
+<div class="bg-info tab-content">
 
    <div id="menu0" class="tab-pane fade in active">
       <!-- Users grid -->
@@ -140,7 +140,7 @@ function getHistory($cfg_file, $accordion_id, $open_first_history_tab = false) {
             <li class="active"><a data-toggle="tab" href="#menu-1-0"><span class="glyphicon glyphicon-heart" aria-hidden="true"></span> Linux</a></li>
             <li><a data-toggle="tab" href="#menu-1-1"><span class="glyphicon glyphicon-th-large" aria-hidden="true"></span> Windows</a></li>
             <li><a data-toggle="tab" href="#menu-1-2"><span class="glyphicon glyphicon-apple" aria-hidden="true"></span> OSX</a></li>
-            <li><a data-toggle="tab" href="#menu-1-3"><span class="glyphicon glyphicon-cog" aria-hidden="true"></span> Server Conf</a></li>
+            <li class="btn-danger"><a data-toggle="tab" href="#menu-1-3"><span class="glyphicon glyphicon-cog" aria-hidden="true"></span> Server Conf</a></li>
 
             <li id="save-config-btn" class="pull-right hidden"><a class="progress-bar-striped" href="javascript:;"><span class="glyphicon glyphicon-save" aria-hidden="true"></span></a></li>
          </ul>
@@ -164,10 +164,11 @@ function getHistory($cfg_file, $accordion_id, $open_first_history_tab = false) {
 
             </div>
             <div id="menu-1-3" class="tab-pane fade">
-
-               <textarea class="form-control" data-config-file="<?= $cfg_file='../vpn/conf/server/server.conf' ?>" name="" id="" cols="30" rows="20"><?= file_get_contents($cfg_file) ?></textarea>
-               <?= getHistory($cfg_file, ++$accId) ?>
-
+               <fieldset>
+                  <legend>Attention! Restart server after changes!</legend>
+                  <textarea class="alert-danger form-control" data-config-file="<?= $cfg_file='../vpn/conf/server/server.conf' ?>" name="" id="" cols="30" rows="20"><?= file_get_contents($cfg_file) ?></textarea>
+                  <?= getHistory($cfg_file, ++$accId) ?>
+               </fieldset>
             </div>
          </div>