+++ install.sh
@@ -6,7 +6,7 @@
 # Original Script from: https://github.com/Chocobozzz/OpenVPN-Admin
 # Parts of the programming from pi-hole were used as templates.
 #
-# changes (c) by Wutze 2020 Version 0.6
+# changes (c) by Wutze 2020 Version 0.7
 #
 # Twitter -> @HuWutze
 
@@ -48,7 +48,9 @@ c=$(( columns / 2 ))
 # Unless the screen is tiny
 r=$(( r < 20 ? 20 : r ))
 c=$(( c < 70 ? 70 : c ))
+h=$(( r - 7 ))
 
+# The script is part of a larger script collection, so this entry exists.
 # If the color table file exists
 if [[ -f "${coltable}" ]]; then
   # source it
@@ -68,46 +70,6 @@ else
 	OVER="\\r\\033[K"
 fi
 
-## Intro with colored Logo
-intro(){
-  NOW=$(date +"%Y")
-	echo -e "${COL_LIGHT_RED}
-■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
-${COL_BLUE}        ◢■◤
-      ◢■◤
-    ◢■◤  ${COL_LIGHT_RED} O P E N V P N - ${COL_NC}W E B A D M I N${COL_LIGHT_RED} - S E R V E R${COL_BLUE}
-  ◢■◤                         【ツ】 © 10.000BC - ${NOW}
-◢■■■■■■■■■■■■■■■■■■■■◤             ${COL_LIGHT_RED}L   I   N   U   X
-■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■${COL_NC}
-"
-}
-
-# read config.conf
-# you must copy config.conf.example to config.conf and edit this file
-if [[ -f "${config}" ]]; then
-  # source it
-  source ${config}
-# Otherwise,
-else
-  echo "Missing configuration"
-  echo "you must copy config.conf.example to config.conf and edit this file"
-  echo "cp config.conf.example config.conf"
-  echo "to edit this file: nano config.conf"
-  exit
-fi
-
-#
-#  name: print_help
-#  @param none
-#  @return message if no options have been specified and exit script
-#  
-print_help () {
-  echo -e "./install.sh www_basedir user group"
-  echo -e "\tbase_dir: The place where the web application will be put in"
-  echo -e "\tuser:     User of the web application"
-  echo -e "\tgroup:    Group of the web application"
-}
-
 #
 #  description: intercepts errors and displays them as messages
 #  name: control_box
@@ -123,7 +85,19 @@ control_box(){
       exit
   fi
 }
-
+#
+#  description: Intercept and display errors
+#  name: control_script
+#  @param $?
+#  @return continue script or or exit when error with exit 100
+#  
+control_script(){
+  if [ ! $? -eq 0 ]
+  then
+  print_out 0 "Error ${1}"
+  exit 100
+  fi
+}
 #
 #  description: formats the notes and messages in an appealing form
 #  name: print_out
@@ -149,18 +123,59 @@ print_out(){
     ;;
   esac
 }
+# Split System-Variable $LANG
+var1=${LANG%.*}
+## Select Language to install
+var2=$(whiptail --title "Select Language" --menu "Select your language" ${r} ${c} ${h} \
+  "AUTO" " Automatic" \
+  "de_DE" " Deutsch" \
+  "en_EN" " Englisch" \
+  3>&1 1>&2 2>&3)
+RET=$?
+if [ $RET -eq 1 ]; then
+  print_out 0 "Exit select language"
+  exit
+elif [ $RET -eq 0 ]; then
+  case "$var2" in
+    AUTO) source lang/"$var1"
+    ;;
+    de_DE) source "lang/$var2"
+    ;;
+    en_EN) source "lang/$var2"
+    ;;
+    *) source "lang/de_DE"
+    ;;
+  esac
+fi
 
-#
-#  description: Intercept and display errors
-#  name: control_script
-#  @param $?
-#  @return continue script or or exit when error with exit 100
-#  
-control_script(){
-  if [ ! $? -eq 0 ]
-  then
-  print_out 0 "Error ${1}"
-  exit 100
+## Intro with colored Logo
+intro(){
+  NOW=$(date +"%Y")
+	echo -e "${COL_LIGHT_RED}
+■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
+${COL_BLUE}        ◢■◤
+      ◢■◤
+    ◢■◤  ${COL_LIGHT_RED} O P E N V P N - ${COL_NC}W E B A D M I N${COL_LIGHT_RED} - S E R V E R${COL_BLUE}
+  ◢■◤                         【ツ】 © 10.000BC - ${NOW}
+◢■■■■■■■■■■■■■■■■■■■■◤             ${COL_LIGHT_RED}L   I   N   U   X
+■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■${COL_NC}
+"
+}
+
+# read config.conf
+# you must copy config.conf.example to config.conf and edit this file
+check_config(){
+  if [[ -f "${config}" ]]; then
+    # source it
+    source ${config}
+  # Otherwise,
+  else
+    echo -e ${COL_LIGHT_RED}$CONFIG01${COL_NC}
+    echo -e $CONFIG02
+    echo -e ${COL_LIGHT_GREEN}$CONFIG03${COL_NC}
+    echo -e $CONFIG04
+    echo -e ${CROSS}" Script "$BREAK
+    exit
   fi
 }
 
@@ -178,79 +193,192 @@ check_user(){
     print_out 1 "${str}"
   else
     print_out 0 "${str}"
-    print_out i "${COL_LIGHT_RED}Script called with non-root privileges${COL_NC}"
-    print_out i "The Installation requires root privileges to install and run"
-    print_out 1 "Installation aborted"
+    print_out i "${COL_LIGHT_RED}${USER01}${COL_NC}"
+    print_out i "${USER02}"
+    print_out 0 "${USER03}"
     exit 1
   fi
 }
 
+do_select(){
+
+#value=("0" "${SELECT00}" on "2" "${SELECT01}" on "3" "${SELECT02}" off)
+#whiptail --title "xx" --checklist "choose" 16 78 10 "${value[@]}"
+# nginx fehlt noch
+	sel=$(whiptail --title "${SELECT0}" --checklist --separate-output "${SELECT1}:" ${r} ${c} ${h} \
+    "1" "${SELECT01} " on \
+    "2" "${SELECT02} " on \
+    "3" "${SELECT03} " on \
+    "4" "${SELECT04} " off \
+    "5" "${SELECT05} " on \
+    "7" "${SELECT07} " on \
+    "9" "${SELECT09} " on \
+    3>&1 1>&2 2>&3)
+#  RET=$?
+  control_box $? "select"
+}
+
+set_autoinstall(){
+  autoinstall1="openvpn php-mysql php-zip php unzip git wget sed curl git net-tools npm nodejs"
+  autoinstall2="npm install -g yarn"
+}
+
+# Future for the progressbar
+go_progress(){
+  apt-get $1
+}
+
+make_mysql(){
+  if [ $mysqlserver ]; then
+    print_out 0 "$FEHLER01 $SELECT02 $FEHLER03 $SELECT03. $ONEONLY"
+    print_out 0 $BREAK
+    exit
+  fi
+  if [ ${1} = 3 ]; then
+    mysqlserver="mariadb-server"
+    sqltest="mysql"
+    installsql="1"
+    echo -e " ${TICK} ${1}" $mysqlserver
+  elif [ ${1} = 4 ]; then
+    mysqlserver="default-mysql-client"
+    sqltest="mysql"
+    installsql=""
+    echo -e " ${TICK} ${1}" $mysqlserver
+  fi
+  #apt-get install mariadb-server -y
+  #echo -e " ${TICK} ${1}"
+}
+
+make_webserver(){
+  if [ $webserver ]; then
+    print_out 0 "$FEHLER01 $SELECT04 $FEHLER03 $SELECT05. $ONEONLY"
+    print_out 0 $BREAK
+    exit
+  fi
+  if [ ${1} = 5 ]; then
+    webserver="apache2"
+    echo -e " ${TICK} ${1}" $webserver
+  elif [ ${1} = 6 ]; then
+    webserver="nginx"
+    echo -e " ${TICK} ${1}" $webserver
+  fi
+    
+}
+
+make_webroot(){
+  www="/srv/www/"
+  echo -e " ${TICK} ${1}" "Set www-root "$www
+}
+
+make_owner(){
+  user="www-data"
+  group="www-data"
+  echo -e " ${TICK} ${1}" "Set permissions: "$user $group
+}
+
+set_config(){
+  cp config.conf.sample config.conf
+}
 #### Start Script with Out- and Inputs
 ## first call funcions
+echo "${ATTENTION}" > info.text
+whiptail --textbox info.text --title "Information" ${r} ${c}
 clear
 intro
-print_out i 'Press enter to continue the VPN-Admin Setup or strg+c to break...'
-print_out r ''
+print_out i "${BEFOR}"
+print_out r 
 check_user
+do_select
 
-# examine the arguments submitted
-if [ "$#" -ne 3 ]; then
-  print_help
-  exit
+while read -r line;
+do #echo "${line}";
+    case $line in
+        1) set_config
+        ;;
+        2) set_autoinstall
+        ;;
+        3|4) make_mysql $line
+        ;;
+        5|6) make_webserver $line
+        ;;
+        7|8) make_webroot $line
+        ;;
+        9|10) make_owner $line
+        ;;
+        *)
+        ;;
+    esac
+done < <(echo "$sel")
+
+check_config
+
+#echo "$webserver $autoinstall1 $mysqlserver -y"
+#echo "$autoinstall2"
+go_progress "update"
+go_progress "install $webserver $autoinstall1 $mysqlserver -y"
+$autoinstall2
+
+# after install mysql-server create mysql-root-pw
+set_mysql_rootpw(){
+	dbpw=$(whiptail --inputbox "${MYSQL01}" ${r} ${c} --title "${MYSQL02}" 3>&1 1>&2 2>&3)
+  control_box $? "set mysql root pw"
+	echo "grant all on *.* to root@localhost identified by '$dbpw' with grant option;" | mysql -u root --password="$dbpw"
+	echo "flush privileges;" | mysql -u root --password="$dbpw"
+}
+
+if [ $installsql ]; then
+  set_mysql_rootpw
 fi
 
 # Ensure there are the prerequisites
 for i in openvpn mysql php yarn node unzip wget sed route; do
   which $i > /dev/null
   if [ "$?" -ne 0 ]; then
-    echo "Miss $i"
+    if [ "$i" = "mysql" ]; then
+      print_out 0 "$INSTMESS"
+      print_out 0 "$BREAK"
+      exit
+    fi
+    print_out 0 "$MISSING ${COL_LIGHT_RED}$i${COL_NC}! $INSTALL"
+    print_out 0 "$BREAK"
     exit
   fi
 done
 
-www=$1
-user=$2
-group=$3
-
 openvpn_admin="$www/openvpn-admin"
-
-# Check the validity of the arguments
-if [ ! -d "$www" ] ||  ! grep -q "$user" "/etc/passwd" || ! grep -q "$group" "/etc/group" ; then
-  print_out 0 "${str}"
-  print_out i "failed Directory, www-user or www-group on your system"
-  exit
-fi
 
 base_path=$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )
 
 ## Message Boxen/Input
-print_out i "give me Input"
-
-ip_server=$(whiptail --inputbox "Server Hostname/IP\nUse the name as the server is to be reached from the Internet!" 8 78 --title "Hostname/IP" 3>&1 1>&2 2>&3)
+## Setup VPN
+ip_server=$(whiptail --inputbox "${SETVPN01}" ${r} ${c} --title "Hostname/IP" 3>&1 1>&2 2>&3)
 control_box $? "Server IP"
-openvpn_proto=$(whiptail --inputbox "OpenVPN protocol (tcp or udp)\nIf you are using a VM with this installation, then select udp:" 8 78 udp --title "Protokoll" 3>&1 1>&2 2>&3)
+openvpn_proto=$(whiptail --inputbox "${SETVPN02}" ${r} ${c} udp --title "Protokoll" 3>&1 1>&2 2>&3)
 control_box $? "VPN Protokoll"
-server_port=$(whiptail --inputbox "OpenVPN Server Port\nDefault Port tcp or udp 1194:" 8 78 1194 --title "Server Port" 3>&1 1>&2 2>&3)
+server_port=$(whiptail --inputbox "${SETVPN03}" ${r} ${c} 1194 --title "Server Port" 3>&1 1>&2 2>&3)
 control_box $? "OpenVPN Port"
 
-db_host=$(whiptail --inputbox "MySQL Host\n(localhost, IP or FQDN)\n\nIf you are using an external database server,\nconfigure it previously so that you can enter a user name and password." 8 78 localhost --title "DB Host" 3>&1 1>&2 2>&3)
+## Setup Database-Server
+db_host=$(whiptail --inputbox "${SETVPN04}" ${r} ${c} localhost --title "DB Host" 3>&1 1>&2 2>&3)
 control_box $? "DB-Host"
+db_name=$(whiptail --inputbox "${SETVPN10}" ${r} ${c} openvpnadmin --title "DB Name" 3>&1 1>&2 2>&3)
+control_box $? "DB-Name"
 
 ## If you are using an external database server
 ## configure it previously so that you can enter a user name and password.
 if [ "$db_host" == localhost ]; then
-  mysql_root_pass=$(whiptail --inputbox "MySQL Root Password\n(The password must not be empty. Please configure this before!)" 8 78 --title "DB Root PW" 3>&1 1>&2 2>&3)
+  mysql_root_pass=$(whiptail --inputbox "${SETVPN05}" ${r} ${c} --title "DB Root PW" 3>&1 1>&2 2>&3)
   control_box $? "Root PW"
 fi
-
-mysql_user=$(whiptail --inputbox "MySQL Username for OpenVPN Database" 8 78 --title "User DB Name" 3>&1 1>&2 2>&3)
+mysql_user=$(whiptail --inputbox "${SETVPN06}" ${r} ${c} --title "User DB Name" 3>&1 1>&2 2>&3)
 control_box $? "MySQL Username"
-mysql_user_pass=$(whiptail --inputbox "MySQL Userpassword for OpenVPN Database" 8 78 --title "User DB PW" 3>&1 1>&2 2>&3)
+mysql_user_pass=$(whiptail --inputbox "${SETVPN07}" ${r} ${c} --title "User DB PW" 3>&1 1>&2 2>&3)
 control_box $? "MySQL User PW"
 
-admin_user=$(whiptail --inputbox "Admin Username for Webfrontend OpenVPN-Admin" 8 78 --title "Web-Admin Name" 3>&1 1>&2 2>&3)
+## Setup Webfrontend
+admin_user=$(whiptail --inputbox "${SETVPN08}" ${r} ${c} --title "Web-Admin Name" 3>&1 1>&2 2>&3)
 control_box $? "Web Admin User"
-admin_user_pass=$(whiptail --inputbox "Admin Userpassword for Webfrontend OpenVPN-Admin" 8 78 --title "Web-Admin PW" 3>&1 1>&2 2>&3)
+admin_user_pass=$(whiptail --inputbox "${SETVPN09}" ${r} ${c} --title "Web-Admin PW" 3>&1 1>&2 2>&3)
 control_box $? "Web Admin PW"
 
 #  
@@ -277,12 +405,12 @@ set_mysql(){
 }
 
 if [ "$db_host" == localhost ]; then
-  set_mysql openvpnadmin $mysql_user $mysql_user_pass
+  set_mysql $db_name $mysql_user $mysql_user_pass
 fi
 
-mysql -h $db_host -u $mysql_user --password=$mysql_user_pass openvpnadmin < sql/vpnadmin.dump
+mysql -h $db_host -u $mysql_user --password=$mysql_user_pass $db_name < sql/vpnadmin.dump
 control_script "Insert Database Dump"
-mysql -h $db_host -u $mysql_user --password=$mysql_user_pass --database=openvpnadmin -e "INSERT INTO admin (admin_id, admin_pass) VALUES ('${admin_user}', encrypt('${admin_user_pass}'));"
+mysql -h $db_host -u $mysql_user --password=$mysql_user_pass --database=$db_name -e "INSERT INTO admin (admin_id, admin_pass) VALUES ('${admin_user}', encrypt('${admin_user_pass}'));"
 control_script "Insert Webadmin User"
 
 print_out 1 "setting up MySQL OK"
@@ -290,6 +418,8 @@ print_out 1 "setting up MySQL OK"
 print_out i "Creating the certificates"
 
 # Get the rsa keys
+# mal austauschen gegen "neu"
+# https://github.com/OpenVPN/easy-rsa/archive/master.zip
 cd /opt/
 wget "https://github.com/OpenVPN/easy-rsa/releases/download/v3.0.6/EasyRSA-unix-v3.0.6.tgz"
 tar -xaf "EasyRSA-unix-v3.0.6.tgz"
@@ -410,9 +540,12 @@ chmod +x "/etc/openvpn/scripts/"*
 # Configure MySQL in openvpn scripts
 sed -i "s/HOST=''/HOST='$db_host'/" "/etc/openvpn/scripts/config.sh"
 sed -i "s/USER=''/USER='$mysql_user'/" "/etc/openvpn/scripts/config.sh"
-sed -i "s/PASS=''/PASS='$mysql_user_pass'/" "/etc/openvpn/scripts/config.sh"
+escaped=$(echo -n "$mysql_user_pass" | sed 's#\\#\\\\#g;s#&#\\&#g')
+sed -i "s/PASS=''/PASS='${escaped}'/" "/etc/openvpn/scripts/config.sh"
+sed -i "s/DB=''/DB='$db_name'/" "/etc/openvpn/scripts/config.sh"
 
 # Create the directory of the web application
+mkdir $www
 mkdir "$openvpn_admin"
 cp -r "$base_path/"{index.php,package.json,js,include,css} "$openvpn_admin"
 mkdir $www/vpn
@@ -426,7 +559,7 @@ cd "$openvpn_admin"
 # Replace config.php variables
 sed -i "s/\$host = '';/\$host = '$db_host';/" "./include/config.php"
 sed -i "s/\$user = '';/\$user = '$mysql_user';/" "./include/config.php"
-escaped=$(echo -n "$mysql_user_pass" | sed 's#\\#\\\\#g;s#&#\\&#g')
+sed -i "s/\$db   = '';/\$db   = '$db_name';/" "./include/config.php"
 sed -i "s/\$pass = '';/\$pass = '${escaped}';/" "./include/config.php"
 
 # Replace in the client configurations with the ip of the server and openvpn protocol
@@ -459,7 +592,7 @@ chown -R "$user:$group" "$openvpn_admin"
 chown -R "$user:$group" $www/vpn
 chown "$user:$group" $www/vpn/conf/server/server.conf
 
-print_out 1 "Finish Installation OpenVPN-Admin"
-print_out i "Please, finish the installation by configuring your web server (Apache, NGinx...)"
-print_out i "Restart your Server after installation please!"
-print_out d "Please, report any issues here https://github.com/wutze/OpenVPN-Admin"
+print_out 1 "${SETFIN01}"
+print_out i "${SETFIN02}"
+print_out i "${SETFIN03}"
+print_out d "${SETFIN04}"